pokemon selector, file discriminator
Fixed an issue where opening the pokemon selector when the image was missing resulted in the default image being written to the cached pokemon. IE crystal nidorina was cached, then we clear the cache. Next time we open, pokemon selector selects Nidorina, default settings are Red and Blue, but it uses the cached settings which means the default Red and Blue sprite would be downloaded and labeled as the Crystal sprite.
Extensions are now enforced on loading and saving profiles, and on saving counter groups